DSPDC-1110 WIP Port to table fragments.
