+++ hle-survey/transformation/src/main/scala/org/broadinstitute/monster/dap/DogTransformations.scala
@@ -6,19 +6,14 @@ import org.broadinstitute.monster.dogaging.jadeschema.table.HlesDog
 object DogTransformations {
 
   /** Parse all dog-related fields out of a raw RedCap record. */
-  def mapDog(rawRecord: RawRecord): HlesDog = {
-    val dogBase = HlesDog.init(
-      dogId = rawRecord.id,
-      ownerId = rawRecord.getRequired("st_owner_id").toLong
-    )
+  def mapDog(rawRecord: RawRecord): HlesDog = HlesDog(
+    dogId = rawRecord.id,
+    ownerId = rawRecord.getRequired("st_owner_id").toLong,
+    hlesDogStudyStatus = Some(StudyStatusTransformations.mapStudyStatus(rawRecord)),
+    hlesDogDemographics = Some(DemographicsTransformations.mapDemographics(rawRecord)),
+    hlesDogResidences = Some(DogResidenceTransformations.mapDogResidences(rawRecord)),
+    hlesDogBehavior = Some(BehaviorTransformations.mapBehavior(rawRecord)),
+    hlesDogHealthSummary = Some(HealthStatusTransformations.mapHealthSummary(rawRecord))
+  )
 
-    val transformations = List(
-      StudyStatusTransformations.mapStudyStatus _,
-      DemographicsTransformations.mapDemographics _,
-      BehaviorTransformations.mapBehavior _,
-      HealthStatusTransformations.mapHealthStatus _
-    )
-
-    transformations.foldLeft(dogBase)((acc, f) => f(rawRecord, acc))
-  }
 }
