+++ hle-survey/transformation/src/main/scala/org/broadinstitute/monster/dap/dog/HealthStatusTransformations.scala
@@ -1,7 +1,7 @@
 package org.broadinstitute.monster.dap.dog
 
 import org.broadinstitute.monster.dap.RawRecord
-import org.broadinstitute.monster.dogaging.jadeschema.table.HlesDog
+import org.broadinstitute.monster.dogaging.jadeschema.fragment.HlesDogHealthSummary
 
 object HealthStatusTransformations {
 
@@ -9,21 +9,21 @@ object HealthStatusTransformations {
     * Parse all high-level health fields out of a raw RedCap record,
     * injecting them into a partially-modeled dog record.
     */
-  def mapHealthStatus(rawRecord: RawRecord, dog: HlesDog): HlesDog = {
+  def mapHealthSummary(rawRecord: RawRecord): HlesDogHealthSummary = {
     val transformations = List(
-      mapHealthSummary _,
+      mapHighLevelFields _,
       mapConditions _,
       mapAltCare _
     )
 
-    transformations.foldLeft(dog)((acc, f) => f(rawRecord, acc))
+    transformations.foldLeft(HlesDogHealthSummary.init())((acc, f) => f(rawRecord, acc))
   }
 
   /**
     * Parse summary health fields out of a raw RedCap record,
     * injecting them into a partially-modeled dog record.
     */
-  def mapHealthSummary(rawRecord: RawRecord, dog: HlesDog): HlesDog = {
+  def mapHighLevelFields(rawRecord: RawRecord, dog: HlesDogHealthSummary): HlesDogHealthSummary = {
     val recentHospitalization = rawRecord.getOptionalBoolean("hs_hosp_yn")
     val hospitalizationReason = recentHospitalization.flatMap {
       if (_) rawRecord.getOptionalNumber("hs_hosp_why") else None
@@ -53,7 +53,7 @@ object HealthStatusTransformations {
     * Parse the Y/N flags for condition types out of a raw RedCap record,
     * injecting them into a partially-modeled dog record as recoded enums.
     */
-  def mapConditions(rawRecord: RawRecord, dog: HlesDog): HlesDog = {
+  def mapConditions(rawRecord: RawRecord, dog: HlesDogHealthSummary): HlesDogHealthSummary = {
 
     def code(congenitalField: Option[String], diagnosedField: Option[String]): Option[Long] = {
       val congenital = congenitalField.flatMap(rawRecord.getOptionalBoolean)
@@ -100,7 +100,7 @@ object HealthStatusTransformations {
     * Parse alternative-care info out of a raw RedCap record,
     * injecting the data into a partially-modeled dog record as "unrolled" booleans.
     */
-  def mapAltCare(rawRecord: RawRecord, dog: HlesDog): HlesDog = {
+  def mapAltCare(rawRecord: RawRecord, dog: HlesDogHealthSummary): HlesDogHealthSummary = {
     val altCareMethods = rawRecord.get("hs_other_health_care")
     val otherAltCare = altCareMethods.map(_.contains("98"))
