+++ hle-survey/transformation/src/main/scala/org/broadinstitute/monster/dap/dog/StudyStatusTransformations.scala
@@ -3,7 +3,7 @@ package org.broadinstitute.monster.dap.dog
 import java.time.LocalDate
 
 import org.broadinstitute.monster.dap.RawRecord
-import org.broadinstitute.monster.dogaging.jadeschema.table.HlesDog
+import org.broadinstitute.monster.dogaging.jadeschema.fragment.HlesDogStudyStatus
 
 object StudyStatusTransformations {
 
@@ -11,8 +11,8 @@ object StudyStatusTransformations {
     * Parse all study-status-related fields out of a raw RedCap record,
     * injecting them into a partially-modeled dog record.
     */
-  def mapStudyStatus(rawRecord: RawRecord, dog: HlesDog): HlesDog =
-    dog.copy(
+  def mapStudyStatus(rawRecord: RawRecord): HlesDogStudyStatus =
+    HlesDogStudyStatus(
       stVipOrStaff = rawRecord.getOptionalNumber("st_vip_or_staff"),
       stBatchLabel = rawRecord.getOptional("st_batch_label"),
       stPortalInvitationDate = rawRecord.getOptionalDate("st_invite_to_portal"),
