+++ schema/src/main/jade-fragments/hles_dog_study_status.fragment.json
@@ -1,0 +1,25 @@
+{
+  "name": "hles_dog_study_status",
+  "columns": [
+    {
+      "name": "st_vip_or_staff",
+      "datatype": "integer"
+    },
+    {
+      "name": "st_batch_label",
+      "datatype": "string"
+    },
+    {
+      "name": "st_portal_invitation_date",
+      "datatype": "date"
+    },
+    {
+      "name": "st_portal_account_creation_date",
+      "datatype": "date"
+    },
+    {
+      "name": "st_hles_completion_date",
+      "datatype": "date"
+    }
+  ]
+}
