--- schema/src/main/jade-structs/hles_dog_residence.struct.json
@@ -1,13 +1,0 @@
-{
-  "name": "hles_dog_residence",
-  "fields": [
-    {
-      "name": "state",
-      "datatype": "string"
-    },
-    {
-      "name": "weeks_at_location",
-      "datatype": "integer"
-    }
-  ]
-}
